Resolve ambiguity of size_t across architectures
git-svn-id: https://svn.smr.arpa.emr.it/arkimet/arkimet/trunk@2435 9b5092fc-ae44-42f6-b796-0d08dad2e17b

+++ arki/emitter.h
@@ -54,8 +54,8 @@ public:
     void add(const char* val) { add_string(val); }
     void add(double val) { add_double(val); }
     void add(int val) { add_int(val); }
-    void add(unsigned val) { add_int(val); }
-    void add(size_t val) { add_int(val); }
+    void add(unsigned int val) { add_int(val); }
+    void add(unsigned long long int val) { add_int(val); }
     void add(long long int val) { add_int(val); }
     void add(bool val) { add_bool(val); }
